Split quarry item-output routing into QuarryOutput
