+++ common/src/main/java/com/logistics/automation/laserquarry/entity/QuarryContext.java
@@ -26,6 +26,8 @@ public interface QuarryContext {
 
     ArmController arm();
 
+    QuarryOutput output();
+
     int breakingEntityId();
 
     int levelMinY();
